requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
uwsgi==2.0.21
drf-spectacular==0.26.2